feat: add user backend

+++ backend/chain/interface.js
@@ -4,11 +4,11 @@ import hash from "./util/hash.js";
 
 export default {
     /* READ */
-    async getAllChains(page = 1, limit = 10) {
+    async getAllChains(userId, page = 1, limit = 10) {
         // Get all chains paginated
         const skip = (page - 1) * limit;
         const [data, total] = await Promise.all([
-            Chain.find({}).skip(skip).limit(limit).sort({ createdAt: -1 }),
+            Chain.find({ userId: userId }).skip(skip).limit(limit).sort({ createdAt: -1 }),
             Chain.countDocuments({})
         ]);
         return { data, total, page, totalPages: Math.ceil(total / limit) };
@@ -19,21 +19,31 @@ export default {
         return await Chain.findById(id);
     },
 
+    async getChainByUser(userId, page = 1, limit = 10) {
+        // Get all chains by a specific user paginated
+        const skip = (page - 1) * limit;
+        const [data, total] = await Promise.all([
+            Chain.find({ userId: userId }).skip(skip).limit(limit).sort({ createdAt: -1 }),
+            Chain.countDocuments({})
+        ]);
+        return { data, total, page, totalPages: Math.ceil(total / limit) };
+    },
+
     /* CREATE */
-    async createChain(name) {
+    async createChain(userId, name) {
         // Create an empty chain
-        return await Chain.create({ name, steps: [], hash: hash([]) });
+        return await Chain.create({ name, userId, steps: [], hash: hash([]) });
     },
 
     /* DELETE */
-    async deleteChain(id) {
+    async deleteChain(userId, id) {
         // Delete chain along with all its execution history
-        await Execution.deleteMany({ chainId: id });
-        return await Chain.findByIdAndDelete(id);
+        await Execution.deleteMany({ userId: id, chainId: id });
+        return await Chain.findOneAndDelete({ _id: id, userId: userId });
     },
 
     /* UPDATE */
-    async updateChain(id, { name, steps }) {    
+    async updateChain(userId, id, { name, steps }) {    
         const updateData  = {};
         
         if (name) updateData.name = name;
@@ -42,8 +52,8 @@ export default {
             updateData.hash = hash([...steps, id]);
         }
 
-        return await Chain.findByIdAndUpdate(
-            id,
+        return await Chain.findOneAndUpdate(
+            { _id: id, userId: userId },
             { $set: updateData },
             { returnDocument: 'after', runValidators: true }
         );
